backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==43.0.3

# Redis
redis==5.0.1

# OpenAI
openai==1.59.5

# Code Parsing - Multi-language support
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-typescript==0.23.2
tree-sitter-java==0.23.4
tree-sitter-c==0.23.1
tree-sitter-cpp==0.23.1
tree-sitter-go==0.23.1
tree-sitter-rust==0.23.0

# Validation
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
requests==2.32.3

# WebSocket
websockets==12.0
